+++ dockci/models/project.py
@@ -25,20 +25,6 @@ from dockci.util import is_yaml_file, is_git_ancestor
 DOCKER_REPO_RE = re.compile(r'[a-z0-9-_.]+')
 
 
-def all_projects():
-    """
-    Get the list of projects
-    """
-    try:
-        for path in Project.data_dir_path().listdir():
-            if is_yaml_file(path):
-                project = Project(path.purebasename)
-                yield project
-
-    except py.error.ENOENT:
-        return
-
-
 class Project(Model):  # pylint:disable=too-few-public-methods
     """
     A project, representing a container to be built
@@ -288,6 +274,8 @@ class Project(Model):  # pylint:disable=too-few-public-methods
     slug = None
     repo = LoadOnAccess(default=lambda _: '')
     name = LoadOnAccess(default=lambda _: '')
+    utility = LoadOnAccess(default=False, input_transform=bool, index=True)
+
     # TODO encrypt decrypt sensitive data etc..
     hipchat_api_token = LoadOnAccess(default=lambda _: '')
     hipchat_room = LoadOnAccess(default=lambda _: '')